+++ src/shared/logger.js
@@ -17,7 +17,7 @@ const levels = {
 const error = (m) => bold(getColor('red')(`☠ ${m}`))
 const warn = (m) => getColor('orangered')(`⚠ ${m}`)
 const success = getColor('amber')
-const successv = getColor('amber')
+const successv = (m) => getColor('amber')(`⟐ ${m}`)
 const info = getColor('gray')
 const help = getColor('dodgerblue')
 const verbose = getColor('plum')
@@ -57,7 +57,7 @@ function formatMessage (level, message) {
     case 'success':
       return success(formattedMessage)
     case 'successv': // success with 'version'
-      return successv(`[${currentName}@${currentVersion}] ${formattedMessage}`)
+      return successv(`${formattedMessage} · ${currentName}@${currentVersion}`)
     // info
     case 'info':
       return info(formattedMessage)