+++ tests/e2e/run.test.js
@@ -63,7 +63,7 @@ t.test('#run', ct => {
 
   const command = `${node} index.js`
   ct.equal(execShell(`${node} index.js`).stdout, 'Hello undefined')
-  ct.equal(execShell(`${dotenvx} run -- ${command}`).stdout, `[dotenvx@${version}] injecting env (1) from .env\nHello World`)
+  ct.equal(execShell(`${dotenvx} run -- ${command}`).stdout, `⟐ injecting env (1) from .env · dotenvx@${version}\nHello World`)
   ct.equal(execShell(`${dotenvx} run --quiet -- ${command}`).stdout, 'Hello World') // --quiet
   ct.equal(execShell(`${dotenvx} run --debug -- ${command}`).stdout, `Setting log level to debug
 process command [${node} index.js]
@@ -72,7 +72,7 @@ loading env from .env (${tempDir}/.env)
 {"HELLO":"World"}
 HELLO set
 HELLO set to World
-[dotenvx@${version}] injecting env (1) from .env
+⟐ injecting env (1) from .env · dotenvx@${version}
 executing process command [${node} index.js]
 expanding process command to [${node} index.js]
 Hello World`) // --debug
@@ -88,7 +88,7 @@ t.test('#run - multiple .env files', ct => {
   `)
 
   const command = `${node} index.js`
-  ct.equal(execShell(`${dotenvx} run -f .env.local -f .env -- ${command}`).stdout, `[dotenvx@${version}] injecting env (1) from .env.local, .env\nHello local`)
+  ct.equal(execShell(`${dotenvx} run -f .env.local -f .env -- ${command}`).stdout, `⟐ injecting env (1) from .env.local, .env · dotenvx@${version}\nHello local`)
   ct.equal(execShell(`${dotenvx} run -f .env.local -f .env --quiet -- ${command}`).stdout, 'Hello local') // --quiet
   ct.equal(execShell(`${dotenvx} run -f .env.local -f .env --debug -- ${command}`).stdout, `Setting log level to debug
 process command [${node} index.js]
@@ -101,7 +101,7 @@ loading env from .env (${tempDir}/.env)
 {"HELLO":"local"}
 HELLO pre-exists (protip: use --overload to override)
 HELLO pre-exists as local (protip: use --overload to override)
-[dotenvx@${version}] injecting env (1) from .env.local, .env
+⟐ injecting env (1) from .env.local, .env · dotenvx@${version}
 executing process command [${node} index.js]
 expanding process command to [${node} index.js]
 Hello local`) // --debug
@@ -117,7 +117,7 @@ t.test('#run - multiple .env files --overload', ct => {
   `)
 
   const command = `${node} index.js`
-  ct.equal(execShell(`${dotenvx} run -f .env.local -f .env --overload -- ${command}`).stdout, `[dotenvx@${version}] injecting env (1) from .env.local, .env\nHello World`)
+  ct.equal(execShell(`${dotenvx} run -f .env.local -f .env --overload -- ${command}`).stdout, `⟐ injecting env (1) from .env.local, .env · dotenvx@${version}\nHello World`)
   ct.equal(execShell(`${dotenvx} run -f .env.local -f .env --overload --quiet -- ${command}`).stdout, 'Hello World') // --quiet
   ct.equal(execShell(`${dotenvx} run -f .env.local -f .env --overload --debug -- ${command}`).stdout, `Setting log level to debug
 process command [${node} index.js]
@@ -130,7 +130,7 @@ loading env from .env (${tempDir}/.env)
 {"HELLO":"World"}
 HELLO set
 HELLO set to World
-[dotenvx@${version}] injecting env (1) from .env.local, .env
+⟐ injecting env (1) from .env.local, .env · dotenvx@${version}
 executing process command [${node} index.js]
 expanding process command to [${node} index.js]
 Hello World`) // --debug
@@ -172,7 +172,7 @@ t.test('#run - --env', ct => {
   `)
 
   const command = `${node} index.js`
-  ct.equal(execShell(`${dotenvx} run --env HELLO=String -f .env -- ${command}`).stdout, `[dotenvx@${version}] injecting env (1) from .env, and --env flag\nHello String`)
+  ct.equal(execShell(`${dotenvx} run --env HELLO=String -f .env -- ${command}`).stdout, `⟐ injecting env (1) from .env, and --env flag · dotenvx@${version}\nHello String`)
   ct.equal(execShell(`${dotenvx} run --env HELLO=String -f .env --quiet -- ${command}`).stdout, 'Hello String') // --quiet
   ct.equal(execShell(`${dotenvx} run --env HELLO=String -f .env --debug -- ${command}`).stdout, `Setting log level to debug
 process command [${node} index.js]
@@ -185,7 +185,7 @@ loading env from .env (${tempDir}/.env)
 {"HELLO":"String"}
 HELLO pre-exists (protip: use --overload to override)
 HELLO pre-exists as String (protip: use --overload to override)
-[dotenvx@${version}] injecting env (1) from .env, and --env flag
+⟐ injecting env (1) from .env, and --env flag · dotenvx@${version}
 executing process command [${node} index.js]
 expanding process command to [${node} index.js]
 Hello String`) // --debug
@@ -204,7 +204,7 @@ t.test('#run - encrypted .env', ct => {
   const DOTENV_PUBLIC_KEY = parsedEnv.DOTENV_PUBLIC_KEY
 
   const command = `${node} index.js`
-  ct.equal(execShell(`${dotenvx} run -- ${command}`).stdout, `[dotenvx@${version}] injecting env (2) from .env\nHello encrypted`)
+  ct.equal(execShell(`${dotenvx} run -- ${command}`).stdout, `⟐ injecting env (2) from .env · dotenvx@${version}\nHello encrypted`)
   ct.equal(execShell(`${dotenvx} run --quiet -- ${command}`).stdout, 'Hello encrypted') // --quiet
   ct.equal(execShell(`${dotenvx} run --debug -- ${command}`).stdout, `Setting log level to debug
 process command [${node} index.js]
@@ -215,7 +215,7 @@ DOTENV_PUBLIC_KEY set
 DOTENV_PUBLIC_KEY set to ${DOTENV_PUBLIC_KEY}
 HELLO set
 HELLO set to encrypted
-[dotenvx@${version}] injecting env (2) from .env
+⟐ injecting env (2) from .env · dotenvx@${version}
 executing process command [${node} index.js]
 expanding process command to [${node} index.js]
 Hello encrypted`) // --debug
@@ -240,7 +240,7 @@ t.test('#run - encrypted .env with no .env.keys', ct => {
 
   let o = execShell(`${dotenvx} run -- ${command}`)
   ct.equal(o.stderr, '☠ [MISSING_PRIVATE_KEY] could not decrypt HELLO using private key \'DOTENV_PRIVATE_KEY=\'. fix: [https://github.com/dotenvx/dotenvx/issues/464]')
-  ct.equal(o.stdout, `[dotenvx@${version}] injecting env (2) from .env\nHello ${encrypted}`)
+  ct.equal(o.stdout, `⟐ injecting env (2) from .env · dotenvx@${version}\nHello ${encrypted}`)
 
   o = execShell(`${dotenvx} run --quiet -- ${command}`)
   ct.equal(o.stderr, '☠ [MISSING_PRIVATE_KEY] could not decrypt HELLO using private key \'DOTENV_PRIVATE_KEY=\'. fix: [https://github.com/dotenvx/dotenvx/issues/464]')
@@ -257,7 +257,7 @@ DOTENV_PUBLIC_KEY set
 DOTENV_PUBLIC_KEY set to ${DOTENV_PUBLIC_KEY}
 HELLO set
 HELLO set to ${encrypted}
-[dotenvx@${version}] injecting env (2) from .env
+⟐ injecting env (2) from .env · dotenvx@${version}
 executing process command [${node} index.js]
 expanding process command to [${node} index.js]
 Hello ${encrypted}`) // --debug
@@ -281,7 +281,7 @@ t.test('#run - encrypted .env with no .env.keys, with DOTENV_PRIVATE_KEY', ct =>
   process.env.DOTENV_PRIVATE_KEY = DOTENV_PRIVATE_KEY // set already on server
 
   const command = `${node} index.js`
-  ct.equal(execShell(`DOTENV_PRIVATE_KEY=${DOTENV_PRIVATE_KEY} ${dotenvx} run -- ${command}`).stdout, `[dotenvx@${version}] injecting env (2) from .env\nHello encrypted`)
+  ct.equal(execShell(`DOTENV_PRIVATE_KEY=${DOTENV_PRIVATE_KEY} ${dotenvx} run -- ${command}`).stdout, `⟐ injecting env (2) from .env · dotenvx@${version}\nHello encrypted`)
   ct.equal(execShell(`${dotenvx} run --quiet -- ${command}`).stdout, 'Hello encrypted') // --quiet
 
   ct.end()
@@ -303,7 +303,7 @@ t.test('#run - encrypted .env.production with no .env.keys, with DOTENV_PRIVATE_
   process.env.DOTENV_PRIVATE_KEY_PRODUCTION = DOTENV_PRIVATE_KEY_PRODUCTION // set already on server
 
   const command = `${node} index.js`
-  ct.equal(execShell(`DOTENV_PRIVATE_KEY_PRODUCTION=${DOTENV_PRIVATE_KEY_PRODUCTION} ${dotenvx} run -- ${command}`).stdout, `[dotenvx@${version}] injecting env (2) from .env.production\nHello production`)
+  ct.equal(execShell(`DOTENV_PRIVATE_KEY_PRODUCTION=${DOTENV_PRIVATE_KEY_PRODUCTION} ${dotenvx} run -- ${command}`).stdout, `⟐ injecting env (2) from .env.production · dotenvx@${version}\nHello production`)
   ct.equal(execShell(`${dotenvx} run --quiet -- ${command}`).stdout, 'Hello production') // --quiet
 
   ct.end()